Add StrictHostKeyChecking=no to git ssh command
This solves the issue of a "Host key verification failure"
occuring in the Jenkins job when trying to clone the
git repository holding the static content.

It was tried to explicitly add the host's key via ssh-keyscan.
However, this was for an unknown reason not successfull.
The approach of adding a git ssh command is chosen since
otherwise the ssh settings in the Jenkins would need to be
touched and changed, which is not prefered.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>